+++ modules/client/src/addresslistVoting/internal/utils.ts
@@ -126,6 +126,15 @@ export function toAddresslistVotingProposalListItem(
       no: proposal.no ? parseInt(proposal.no) : 0,
       abstain: proposal.abstain ? parseInt(proposal.abstain) : 0,
     },
+    actions: proposal.actions.map(
+      (action: SubgraphAction): DaoAction => {
+        return {
+          data: hexToBytes(action.data),
+          to: action.to,
+          value: BigInt(action.value),
+        };
+      },
+    ),
     votes: proposal.voters.map(
       (voter: SubgraphAddresslistVotingVoterListItem) => {
         return {