+++ modules/client/CHANGELOG.md
@@ -18,6 +18,17 @@ TEMPLATE:
 -->
 
 ## [UPCOMING]
+
+### Added
+
+- `isMember` function to `TokenVotingClient`
+- `isMember` function to `AddresslistVotingClient`
+- `isMember` function to `Client`
+- Add `actions` to `MultisigProposalListItem`
+- Add `actions` to `TokenVotingProposalListItem`
+- Add `actions` to `AddresslistVotingProposalListItem`
+
+## [1.18.2]
 ### Fixed
 
 - Plugin preparations query